requirements.txt
#decord - Does not work on mac
eva-decord
ffmpeg_python
hydra-core
numpy
opencv_python
Pillow
scipy

cmake


#torch
#torchaudio
#torchvision
tqdm


